added forgotpassword

+++ controllers/courses.js
@@ -49,6 +49,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
   if (!bootcamp) {
@@ -58,7 +59,17 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
     );
   }
 
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401
+      )
+    );
+  }
   const course = await Course.create(req.body);
+
   res.status(200).json({
     success: true,
     data: course,
@@ -75,6 +86,16 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
     return next(
       new ErrorResponse(`No bootcamp with the id of ${req.params.id} `),
       404
+    );
+  }
+
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
     );
   }
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
@@ -107,17 +128,17 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
     );
   }
 
-  // Make sure user is bootcamp owner
-  // if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
-  //   return next(
-  //     new ErrorResponse(
-  //       `User with ID that ends with '${req.user.id.slice(
-  //         -6
-  //       )}' is not authorized to delete this course`,
-  //       401
-  //     )
-  //   );
-  // }
+  //Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User with ID that ends with '${req.user.id.slice(
+          -6
+        )}' is not authorized to delete this course`,
+        401
+      )
+    );
+  }
 
   await course.deleteOne();
